+++ rows/plugins/txt.py
@@ -17,7 +17,8 @@
 
 from __future__ import unicode_literals
 
-from rows.plugins.utils import create_table, get_filename_and_fobj, serialize
+from rows.plugins.utils import (create_table, export_data,
+                                get_filename_and_fobj, serialize)
 
 DASH, PLUS, PIPE = '-', '+', '|'
 
@@ -44,13 +45,12 @@ def import_from_txt(filename_or_fobj, encoding='utf-8', *args, **kwargs):
     return create_table(table_rows, meta=meta, *args, **kwargs)
 
 
-def export_to_txt(table, filename_or_fobj, encoding='utf-8', *args, **kwargs):
+def export_to_txt(table, filename_or_fobj=None, encoding='utf-8', *args, **kwargs):
     # TODO: should be able to change DASH, PLUS and PIPE
     # TODO: will work only if table.fields is OrderedDict
     # TODO: should use fobj? What about creating a method like json.dumps?
 
     kwargs['encoding'] = encoding
-    _, fobj = get_filename_and_fobj(filename_or_fobj, mode='wb')
     serialized_table = serialize(table, *args, **kwargs)
     field_names = serialized_table.next()
     table_rows = list(serialized_table)
@@ -68,9 +68,6 @@ def export_to_txt(table, filename_or_fobj, encoding='utf-8', *args, **kwargs):
         row_data = ' {} '.format(PIPE).join(values)
         result.append('{} {} {}'.format(PIPE, row_data, PIPE))
     result.extend([split_line, ''])
-
     data = '\n'.join(result).encode(encoding)
 
-    fobj.write(data)
-    fobj.flush()
-    return fobj
+    return export_data(filename_or_fobj, data)